+++ c8y_api/model/inventory.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from typing import Any, Generator, List
 
-from c8y_api.model._base import CumulocityResource
+from c8y_api.model._base import CumulocityResource, get_all_by_path
 from c8y_api.model._util import _QueryUtil
 from c8y_api.model.managedobjects import ManagedObjectUtil, ManagedObject, Device, Availability, DeviceGroup
 
@@ -61,6 +61,7 @@ class Inventory(CumulocityResource):
             reverse: bool = None,
             limit: int = None,
             page_size: int = 1000,
+            as_tuples: list[str] | dict[str, Any] = None,
             **kwargs) -> List[ManagedObject]:
         """ Query the database for managed objects and return the results
         as list.
@@ -93,6 +94,7 @@ class Inventory(CumulocityResource):
             reverse=reverse,
             limit=limit,
             page_size=page_size,
+            as_tuples=as_tuples,
             **kwargs))
 
     def get_count(
@@ -154,6 +156,7 @@ class Inventory(CumulocityResource):
             limit: int = None,
             page_size: int = 1000,
             page_number: int = None,
+            as_tuples: list[str] | dict[str, Any] = None,
             **kwargs) -> Generator[ManagedObject]:
         """ Query the database for managed objects and iterate over the
         results.
@@ -205,6 +208,10 @@ class Inventory(CumulocityResource):
                 parsed in one chunk). This is a performance related setting.
             page_number (int): Pull a specific page; this effectively disables
                 automatic follow-up page retrieval.
+            as_tuples: (list[str] or dict[str, Any]):  Don't parse ManagedObjects,
+                but extract the values at certain JSON paths as tuples; If the path
+                is not defined in a result, None is used; Specify a dictionary to
+                define proper default values for each path.
 
         Returns:
             Generator for ManagedObject instances
@@ -234,6 +241,7 @@ class Inventory(CumulocityResource):
             limit=limit,
             page_size=page_size,
             page_number=page_number,
+            as_tuples=as_tuples,
             **kwargs)
 
     def _prepare_inventory_query(
@@ -336,10 +344,14 @@ class Inventory(CumulocityResource):
 
         return {query_key: query, **kwargs}
 
-    def _select(self, parse_fun, device_mode: bool, page_number, limit, **kwargs) -> Generator[Any]:
+    def _select(self, parse_fun, device_mode: bool, page_number, limit, as_tuples, **kwargs) -> Generator[Any]:
         """Generic select function to be used by derived classes as well."""
         base_query = self._prepare_inventory_query(device_mode, **kwargs)
-        return super()._iterate(base_query, page_number, limit, parse_fun)
+        return super()._iterate(
+            base_query,
+            page_number,
+            limit,
+            parse_fun if not as_tuples else (lambda x: get_all_by_path(x, as_tuples)))
 
     def create(self, *objects: ManagedObject):
         """Create managed objects within the database.
@@ -489,6 +501,7 @@ class DeviceInventory(Inventory):
             limit: int = None,
             page_size: int = 100,
             page_number: int = None,
+            as_tuples: list[str] | dict[str, Any] = None,
             **kwargs,) -> Generator[Device]:
         # pylint: disable=arguments-differ, arguments-renamed
         """ Query the database for devices and iterate over the results.
@@ -543,6 +556,10 @@ class DeviceInventory(Inventory):
                 parsed in one chunk). This is a performance related setting.
             page_number (int): Pull a specific page; this effectively disables
                 automatic follow-up page retrieval.
+            as_tuples: (list[str] or dict[str, Any]):  Don't parse Device objects,
+                but extract the values at certain JSON paths as tuples; If the path
+                is not defined in a result, None is used; Specify a dictionary to
+                define proper default values for each path.
 
         Returns:
             Generator for Device objects
@@ -571,6 +588,7 @@ class DeviceInventory(Inventory):
             limit=limit,
             page_size=page_size,
             page_number=page_number,
+            as_tuples=as_tuples,
             **kwargs)
 
     def get_all(  # noqa (changed signature)
@@ -596,6 +614,7 @@ class DeviceInventory(Inventory):
             limit: int = None,
             page_size: int = 100,
             page_number: int = None,
+            as_tuples: list[str] | dict[str, Any] = None,
             **kwargs) -> List[Device]:
         # pylint: disable=arguments-differ, arguments-renamed
         """ Query the database for devices and return the results as list.
@@ -628,6 +647,7 @@ class DeviceInventory(Inventory):
             limit=limit,
             page_size=page_size,
             page_number=page_number,
+            as_tuples=as_tuples,
             **kwargs))
 
     def get_count(  # noqa (changed signature)
@@ -734,6 +754,7 @@ class DeviceGroupInventory(Inventory):
             limit: int = None,
             page_size: int = 100,
             page_number: int = None,
+            as_tuples: list[str] | dict[str, Any] = None,
             **kwargs) -> Generator[DeviceGroup]:
         # pylint: disable=arguments-differ, arguments-renamed
         """ Select device groups by various parameters.
@@ -791,6 +812,11 @@ class DeviceGroupInventory(Inventory):
                 parsed in one chunk). This is a performance related setting.
             page_number (int): Pull a specific page; this effectively disables
                 automatic follow-up page retrieval.
+            as_tuples: (list[str] or dict[str, Any]):  Don't parse DeviceGroup
+                objects, but extract the values at certain JSON paths as
+                tuples; If the path is not defined in a result, None is used;
+                Specify a dictionary to define proper default values for each
+                path.
 
         Returns:
             Generator of DeviceGroup instances
@@ -827,6 +853,7 @@ class DeviceGroupInventory(Inventory):
             limit=limit,
             page_size=page_size,
             page_number=page_number,
+            as_tuples=as_tuples,
             **kwargs)
 
     def get_count(  # noqa (changed signature)
@@ -896,7 +923,8 @@ class DeviceGroupInventory(Inventory):
             limit: int = None,
             page_size: int = 100,
             page_number: int = None,
-            **kwargs) -> List[DeviceGroup]:
+            as_tuples: list[str] | dict[str, Any] = None,
+            **kwargs ) -> List[DeviceGroup]:
         # pylint: disable=arguments-differ, arguments-renamed
         """ Select managed objects by various parameters.
 
@@ -927,6 +955,7 @@ class DeviceGroupInventory(Inventory):
             limit=limit,
             page_size=page_size,
             page_number=page_number,
+            as_tuples=as_tuples,
             **kwargs))
 
     def create(self, *groups):